Input block on the Red areas.
This update blocks the input on the red areas because yeah, this tool is  literally meant to do the fill work on the red areas for you lol.

More in-depth:

Added the disabled attribute to all of the red spots on each leg including the addRows function to make the new inserts work with this as well! :)